+++ pkg/storage/measured/blocklist.go
@@ -23,33 +23,34 @@ import (
 )
 
 type measuredBlockListRep struct {
-	rep repository.BlockList
+	rep  repository.BlockList
+	inTx bool
 }
 
 func (m *measuredBlockListRep) UpsertBlockListItem(ctx context.Context, item *blocklistmodel.Item) (err error) {
 	t0 := time.Now()
 	err = m.rep.UpsertBlockListItem(ctx, item)
-	reportOpMetric(upsertOp, time.Since(t0).Seconds(), err == nil)
+	reportOpMetric(upsertOp, time.Since(t0).Seconds(), err == nil, m.inTx)
 	return
 }
 
 func (m *measuredBlockListRep) DeleteBlockListItem(ctx context.Context, item *blocklistmodel.Item) (err error) {
 	t0 := time.Now()
 	err = m.rep.DeleteBlockListItem(ctx, item)
-	reportOpMetric(deleteOp, time.Since(t0).Seconds(), err == nil)
+	reportOpMetric(deleteOp, time.Since(t0).Seconds(), err == nil, m.inTx)
 	return
 }
 
 func (m *measuredBlockListRep) FetchBlockListItems(ctx context.Context, username string) (blockList []*blocklistmodel.Item, err error) {
 	t0 := time.Now()
 	blockList, err = m.rep.FetchBlockListItems(ctx, username)
-	reportOpMetric(fetchOp, time.Since(t0).Seconds(), err == nil)
+	reportOpMetric(fetchOp, time.Since(t0).Seconds(), err == nil, m.inTx)
 	return
 }
 
 func (m *measuredBlockListRep) DeleteBlockListItems(ctx context.Context, username string) (err error) {
 	t0 := time.Now()
 	err = m.rep.DeleteBlockListItems(ctx, username)
-	reportOpMetric(deleteOp, time.Since(t0).Seconds(), err == nil)
+	reportOpMetric(deleteOp, time.Since(t0).Seconds(), err == nil, m.inTx)
 	return
 }