+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 ## jackal - main / unreleased
 
 * [ENHANCEMENT] Added memory ballast. #198
+* [CHANGE] Introduced measured repository transaction type. #200 
 * [BUGFIX] Fix S2S db key check when nop KV is used. #199